+++ src/main/java/kr/sesacjava/swimtutor/common/requestLog/MethodType.java
@@ -1,4 +1,4 @@
-package kr.sesacjava.swimtutor.leveltest.entity;
+package kr.sesacjava.swimtutor.common.requestLog;
 
 public enum MethodType {
     GET, POST, PUT, DELETE
