+++ src/main/java/kr/sesacjava/swimtutor/leveltest/repository/RequestLogRepository.java
@@ -1,6 +1,6 @@
 package kr.sesacjava.swimtutor.leveltest.repository;
 
-import kr.sesacjava.swimtutor.leveltest.entity.RequestLog;
+import kr.sesacjava.swimtutor.common.requestLog.RequestLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RequestLogRepository extends JpaRepository<RequestLog, Integer> {
